Added support for Notes for Tribe favor and displaying them on the selection board
Added initial support for Tribe favor reward notes and to display them on the main ToTA board. Currently using fixed values for size and offset to draw the text box. Either dynamic relative values should be used or a set of sliders for size and offset should be added for these values so that they work across all resolutions correctly. Check the code source comments for more details.